+++ sdk/python/feast/infra/compute_engines/spark/compute.py
@@ -1,7 +1,7 @@
 import pyarrow as pa
 
 from feast.infra.compute_engines.base import ComputeEngine, HistoricalRetrievalTask
-from feast.infra.compute_engines.dag.model import ExecutionContext
+from feast.infra.compute_engines.dag.context import ExecutionContext
 from feast.infra.compute_engines.spark.spark_dag_builder import SparkDAGBuilder
 from feast.infra.materialization.batch_materialization_engine import (
     MaterializationJob,
